# internal/api/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [3]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'c': // Prefix: "config"

				if l := len("config"); len(elem) >= l && elem[0:l] == "config" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleConfigGetRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleConfigPostRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET,POST")
					}

					return
				}

			case 't': // Prefix: "tickets"

				if l := len("tickets"); len(elem) >= l && elem[0:l] == "tickets" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleTicketsGetRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleTicketsPostRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET,POST")
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "ticketId"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch r.Method {
						case "DELETE":
							s.handleTicketsTicketIdDeleteRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "GET":
							s.handleTicketsTicketIdGetRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "PATCH":
							s.handleTicketsTicketIdPatchRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "DELETE,GET,PATCH")
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 'a': // Prefix: "ai/generate"

							if l := len("ai/generate"); len(elem) >= l && elem[0:l] == "ai/generate" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "POST":
									s.handleTicketsTicketIdAiGeneratePostRequest([1]string{
										args[0],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, "POST")
								}

								return
							}

						case 'n': // Prefix: "notes"

							if l := len("notes"); len(elem) >= l && elem[0:l] == "notes" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								switch r.Method {
								case "POST":
									s.handleTicketsTicketIdNotesPostRequest([1]string{
										args[0],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, "POST")
								}

								return
							}
							switch elem[0] {
							case '/': // Prefix: "/"

								if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
									elem = elem[l:]
								} else {
									break
								}

								// Param: "noteId"
								// Match until "/"
								idx := strings.IndexByte(elem, '/')
								if idx < 0 {
									idx = len(elem)
								}
								args[1] = elem[:idx]
								elem = elem[idx:]

								if len(elem) == 0 {
									switch r.Method {
									case "DELETE":
										s.handleTicketsTicketIdNotesNoteIdDeleteRequest([2]string{
											args[0],
											args[1],
										}, elemIsEscaped, w, r)
									case "PUT":
										s.handleTicketsTicketIdNotesNoteIdPutRequest([2]string{
											args[0],
											args[1],
										}, elemIsEscaped, w, r)
									default:
										s.notAllowed(w, r, "DELETE,PUT")
									}

									return
								}
								switch elem[0] {
								case '/': // Prefix: "/"

									if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
										elem = elem[l:]
									} else {
										break
									}

									if len(elem) == 0 {
										break
									}
									switch elem[0] {
									case 'a': // Prefix: "ai/review"

										if l := len("ai/review"); len(elem) >= l && elem[0:l] == "ai/review" {
											elem = elem[l:]
										} else {
											break
										}

										if len(elem) == 0 {
											// Leaf node.
											switch r.Method {
											case "POST":
												s.handleTicketsTicketIdNotesNoteIdAiReviewPostRequest([2]string{
													args[0],
													args[1],
												}, elemIsEscaped, w, r)
											default:
												s.notAllowed(w, r, "POST")
											}

											return
										}

									case 'r': // Prefix: "reviews"

										if l := len("reviews"); len(elem) >= l && elem[0:l] == "reviews" {
											elem = elem[l:]
										} else {
											break
										}

										if len(elem) == 0 {
											switch r.Method {
											case "POST":
												s.handleCreateReviewRequest([2]string{
													args[0],
													args[1],
												}, elemIsEscaped, w, r)
											default:
												s.notAllowed(w, r, "POST")
											}

											return
										}
										switch elem[0] {
										case '/': // Prefix: "/"

											if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
												elem = elem[l:]
											} else {
												break
											}

											// Param: "reviewId"
											// Leaf parameter, slashes are prohibited
											idx := strings.IndexByte(elem, '/')
											if idx >= 0 {
												break
											}
											args[2] = elem
											elem = ""

											if len(elem) == 0 {
												// Leaf node.
												switch r.Method {
												case "DELETE":
													s.handleDeleteReviewRequest([3]string{
														args[0],
														args[1],
														args[2],
													}, elemIsEscaped, w, r)
												case "PUT":
													s.handleUpdateReviewRequest([3]string{
														args[0],
														args[1],
														args[2],
													}, elemIsEscaped, w, r)
												default:
													s.notAllowed(w, r, "DELETE,PUT")
												}

												return
											}

										}

									}

								}

							}

						}

					}

				}

			case 'u': // Prefix: "users"

				if l := len("users"); len(elem) >= l && elem[0:l] == "users" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleUsersGetRequest([0]string{}, elemIsEscaped, w, r)
					case "PUT":
						s.handleUsersPutRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET,PUT")
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [3]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'c': // Prefix: "config"

				if l := len("config"); len(elem) >= l && elem[0:l] == "config" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = ConfigGetOperation
						r.summary = "設定情報の取得"
						r.operationID = ""
						r.operationGroup = ""
						r.pathPattern = "/config"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = ConfigPostOperation
						r.summary = "設定情報の更新"
						r.operationID = ""
						r.operationGroup = ""
						r.pathPattern = "/config"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 't': // Prefix: "tickets"

				if l := len("tickets"); len(elem) >= l && elem[0:l] == "tickets" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = TicketsGetOperation
						r.summary = "チケット一覧取得"
						r.operationID = ""
						r.operationGroup = ""
						r.pathPattern = "/tickets"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = TicketsPostOperation
						r.summary = "チケット新規作成"
						r.operationID = ""
						r.operationGroup = ""
						r.pathPattern = "/tickets"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "ticketId"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch method {
						case "DELETE":
							r.name = TicketsTicketIdDeleteOperation
							r.summary = "チケット削除"
							r.operationID = ""
							r.operationGroup = ""
							r.pathPattern = "/tickets/{ticketId}"
							r.args = args
							r.count = 1
							return r, true
						case "GET":
							r.name = TicketsTicketIdGetOperation
							r.summary = "チケット詳細取得"
							r.operationID = ""
							r.operationGroup = ""
							r.pathPattern = "/tickets/{ticketId}"
							r.args = args
							r.count = 1
							return r, true
						case "PATCH":
							r.name = TicketsTicketIdPatchOperation
							r.summary = "チケット情報更新"
							r.operationID = ""
							r.operationGroup = ""
							r.pathPattern = "/tickets/{ticketId}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 'a': // Prefix: "ai/generate"

							if l := len("ai/generate"); len(elem) >= l && elem[0:l] == "ai/generate" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "POST":
									r.name = TicketsTicketIdAiGeneratePostOperation
									r.summary = "AIによる返信ドラフト生成 (SSE)"
									r.operationID = ""
									r.operationGroup = ""
									r.pathPattern = "/tickets/{ticketId}/ai/generate"
									r.args = args
									r.count = 1
									return r, true
								default:
									return
								}
							}

						case 'n': // Prefix: "notes"

							if l := len("notes"); len(elem) >= l && elem[0:l] == "notes" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								switch method {
								case "POST":
									r.name = TicketsTicketIdNotesPostOperation
									r.summary = "ノート追加"
									r.operationID = ""
									r.operationGroup = ""
									r.pathPattern = "/tickets/{ticketId}/notes"
									r.args = args
									r.count = 1
									return r, true
								default:
									return
								}
							}
							switch elem[0] {
							case '/': // Prefix: "/"

								if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
									elem = elem[l:]
								} else {
									break
								}

								// Param: "noteId"
								// Match until "/"
								idx := strings.IndexByte(elem, '/')
								if idx < 0 {
									idx = len(elem)
								}
								args[1] = elem[:idx]
								elem = elem[idx:]

								if len(elem) == 0 {
									switch method {
									case "DELETE":
										r.name = TicketsTicketIdNotesNoteIdDeleteOperation
										r.summary = "ノート削除"
										r.operationID = ""
										r.operationGroup = ""
										r.pathPattern = "/tickets/{ticketId}/notes/{noteId}"
										r.args = args
										r.count = 2
										return r, true
									case "PUT":
										r.name = TicketsTicketIdNotesNoteIdPutOperation
										r.summary = "ノート編集"
										r.operationID = ""
										r.operationGroup = ""
										r.pathPattern = "/tickets/{ticketId}/notes/{noteId}"
										r.args = args
										r.count = 2
										return r, true
									default:
										return
									}
								}
								switch elem[0] {
								case '/': // Prefix: "/"

									if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
										elem = elem[l:]
									} else {
										break
									}

									if len(elem) == 0 {
										break
									}
									switch elem[0] {
									case 'a': // Prefix: "ai/review"

										if l := len("ai/review"); len(elem) >= l && elem[0:l] == "ai/review" {
											elem = elem[l:]
										} else {
											break
										}

										if len(elem) == 0 {
											// Leaf node.
											switch method {
											case "POST":
												r.name = TicketsTicketIdNotesNoteIdAiReviewPostOperation
												r.summary = "AIによるノート（下書き）のレビュー (SSE)"
												r.operationID = ""
												r.operationGroup = ""
												r.pathPattern = "/tickets/{ticketId}/notes/{noteId}/ai/review"
												r.args = args
												r.count = 2
												return r, true
											default:
												return
											}
										}

									case 'r': // Prefix: "reviews"

										if l := len("reviews"); len(elem) >= l && elem[0:l] == "reviews" {
											elem = elem[l:]
										} else {
											break
										}

										if len(elem) == 0 {
											switch method {
											case "POST":
												r.name = CreateReviewOperation
												r.summary = "レビュー追加"
												r.operationID = "createReview"
												r.operationGroup = ""
												r.pathPattern = "/tickets/{ticketId}/notes/{noteId}/reviews"
												r.args = args
												r.count = 2
												return r, true
											default:
												return
											}
										}
										switch elem[0] {
										case '/': // Prefix: "/"

											if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
												elem = elem[l:]
											} else {
												break
											}

											// Param: "reviewId"
											// Leaf parameter, slashes are prohibited
											idx := strings.IndexByte(elem, '/')
											if idx >= 0 {
												break
											}
											args[2] = elem
											elem = ""

											if len(elem) == 0 {
												// Leaf node.
												switch method {
												case "DELETE":
													r.name = DeleteReviewOperation
													r.summary = "レビュー取り消し"
													r.operationID = "deleteReview"
													r.operationGroup = ""
													r.pathPattern = "/tickets/{ticketId}/notes/{noteId}/reviews/{reviewId}"
													r.args = args
													r.count = 3
													return r, true
												case "PUT":
													r.name = UpdateReviewOperation
													r.summary = "レビュー修正"
													r.operationID = "updateReview"
													r.operationGroup = ""
													r.pathPattern = "/tickets/{ticketId}/notes/{noteId}/reviews/{reviewId}"
													r.args = args
													r.count = 3
													return r, true
												default:
													return
												}
											}

										}

									}

								}

							}

						}

					}

				}

			case 'u': // Prefix: "users"

				if l := len("users"); len(elem) >= l && elem[0:l] == "users" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = UsersGetOperation
						r.summary = "ユーザー一覧取得"
						r.operationID = ""
						r.operationGroup = ""
						r.pathPattern = "/users"
						r.args = args
						r.count = 0
						return r, true
					case "PUT":
						r.name = UsersPutOperation
						r.summary = "ユーザー情報の同期・更新"
						r.operationID = ""
						r.operationGroup = ""
						r.pathPattern = "/users"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}